requirements.txt
pypdf>=5.0.0
Pillow>=11.0.0
rich>=13.9.0
exifread>=3.0.0
